Moved event counter to refactored
