+++ hawkbit-repository/hawkbit-repository-jpa/pom.xml
@@ -133,12 +133,6 @@
          <artifactId>powermock-api-mockito</artifactId>
          <scope>test</scope>
       </dependency>
-      <dependency>
-         <groupId>org.easytesting</groupId>
-         <artifactId>fest-assert-core</artifactId>
-         <scope>test</scope>
-         
-      </dependency>
    </dependencies>
 
    <build>
