--- hawkbit-repository/hawkbit-repository-jpa/src/test/java/org/eclipse/hawkbit/repository/jpa/utils/ExpectEvent.java
@@ -1,15 +1,0 @@
-package org.eclipse.hawkbit.repository.jpa.utils;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ ElementType.METHOD, ElementType.TYPE })
-public @interface ExpectEvent {
-
-    Class<?> type();
-
-    int count() default 1;
-}
