+++ hawkbit-repository/hawkbit-repository-test/pom.xml
@@ -90,5 +90,10 @@
          <groupId>io.protostuff</groupId>
          <artifactId>protostuff-runtime</artifactId>
       </dependency>
+      <dependency>
+         <groupId>org.easytesting</groupId>
+         <artifactId>fest-assert-core</artifactId>
+         <scope>test</scope>
+      </dependency>
   </dependencies>
 </project>
