+++ hawkbit-repository/hawkbit-repository-test/src/main/java/org/eclipse/hawkbit/repository/test/util/EventCounterListener.java
@@ -1,4 +1,12 @@
-package org.eclipse.hawkbit.repository.jpa.utils;
+/**
+ * Copyright (c) 2015 Bosch Software Innovations GmbH and others.
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+package org.eclipse.hawkbit.repository.test.util;
 
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
@@ -7,6 +15,9 @@ import java.util.concurrent.atomic.AtomicInteger;
 import org.springframework.cloud.bus.event.RemoteApplicationEvent;
 import org.springframework.context.ApplicationListener;
 
+/**
+ * Event counter listener to count events and save the type and count in a map.
+ */
 public class EventCounterListener implements ApplicationListener<RemoteApplicationEvent> {
 
     final ConcurrentMap<Class<?>, AtomicInteger> map;
