+++ hawkbit-repository/hawkbit-repository-test/src/main/java/org/eclipse/hawkbit/repository/test/util/EventVerifier.java
@@ -1,17 +1,27 @@
-package org.eclipse.hawkbit.repository.jpa.utils;
+/**
+ * Copyright (c) 2015 Bosch Software Innovations GmbH and others.
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+package org.eclipse.hawkbit.repository.test.util;
 
 import static org.fest.assertions.api.Assertions.assertThat;
 
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.atomic.AtomicInteger;
 
-import org.eclipse.hawkbit.repository.test.util.TestApplicationContextProvider;
 import org.junit.rules.TestRule;
 import org.junit.runner.Description;
 import org.junit.runners.model.Statement;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.event.ApplicationEventMulticaster;
 
+/**
+ * Test rule to setup and to verify the event count for a method.
+ */
 public class EventVerifier implements TestRule {
 
     EventCounterListener eventCounterListener;
@@ -42,30 +52,33 @@ public class EventVerifier implements TestRule {
 
     private void afterTest(final Description description) throws InterruptedException {
 
-        if (hasEventCountAnnotation(description)) {
-            final ExpectEvent[] expectEvent = getExpectEvent(description);
-            if (expectEvent.length > 0) {
-                final ConcurrentMap<Class<?>, AtomicInteger> eventCounterMap = eventCounterListener
-                        .getEventCounterMap();
-                for (final ExpectEvent e : expectEvent) {
-                    eventCounterMap.putIfAbsent(e.type(), new AtomicInteger(0));
-                    final AtomicInteger atomicInteger = eventCounterMap.get(e.type());
-                    assertThat(atomicInteger.get()).as("Did not receive the expected amount of events form " + e.type())
-                            .isEqualTo(e.count());
-                }
-            }
+        if (!hasEventCountAnnotation(description)) {
+            return;
         }
-        removeEventCounterListener();
+
+        final ExpectEvent[] expectEvent = getExpectEvent(description);
+        if (expectEvent.length > 0) {
+            verifyEventCounts(expectEvent);
+        }
+
+        removeEventCounterListenerIfNecessary();
+    }
+
+    private void verifyEventCounts(final ExpectEvent[] expectEvent) {
+        final ConcurrentMap<Class<?>, AtomicInteger> eventCounterMap = eventCounterListener.getEventCounterMap();
+        for (final ExpectEvent e : expectEvent) {
+            eventCounterMap.putIfAbsent(e.type(), new AtomicInteger(0));
+            final AtomicInteger atomicInteger = eventCounterMap.get(e.type());
+            assertThat(atomicInteger.get()).as("Did not receive the expected amount of events form " + e.type())
+                    .isEqualTo(e.count());
+        }
     }
 
     private boolean hasEventCountAnnotation(final Description description) {
         return description.getAnnotation(EventCounter.class) != null;
     }
 
-    private void removeEventCounterListener() {
-        if (eventCounterListener == null) {
-            return;
-        }
+    private void removeEventCounterListenerIfNecessary() {
         final ConfigurableApplicationContext context = TestApplicationContextProvider.getContext();
         final ApplicationEventMulticaster applicationEventMulticaster = context
                 .getBean(ApplicationEventMulticaster.class);
